# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REST API приложения для тестового задания",
        "description": "API-документация для CRUD организаций, зданий, видов деятельности и аутентификации.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/organizations/in-building": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список всех организаций находящихся в конкретном здании",
                "operationId": "82313ef247dd73cbd53ac56e81d99d53",
                "parameters": [
                    {
                        "name": "building_id",
                        "in": "query",
                        "description": "ID здания",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество результатов на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/organizations/by-activity": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список всех организаций, которые относятся к указанному виду деятельности",
                "operationId": "1724dd38e57cca320df87e74a582f688",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "query",
                        "description": "ID деятельности",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество результатов на странице (по умолчанию 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/organizations/by-radius": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список организаций, которые находятся в заданном радиусе",
                "operationId": "06241e9e0010ab2af68b9af50e38dc36",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус поиска",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество результатов на странице (по умолчанию 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/organizations/by-box": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Cписок организаций, которые находятся в заданной прямоугольной области",
                "operationId": "88069184f006529997dc18e388f60f0c",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lat_2",
                        "in": "query",
                        "description": "Широта 2",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng_2",
                        "in": "query",
                        "description": "Долгота 2",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество результатов на странице (по умолчанию 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/organizations/by-name": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организации по названию",
                "operationId": "cf8660b01345911f2cf0e51f31cbbf45",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество результатов на странице (по умолчанию 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить информацию об организации по ID",
                "operationId": "c43fd8909d9758788d18f58819842e58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об организации",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Обновить данные об организации",
                "operationId": "46a60849ec79411225ea3ed18dba7bd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "building_id",
                                    "name"
                                ],
                                "properties": {
                                    "building_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "ООО Металстрой"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Организация обновлена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Удалить организацию",
                "operationId": "ea83ebb8503aa55a1a7ec8d026c8d566",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организация удалена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}